fix(workspaces): remove order-dependent import cycle
